+++ tools/admin-console/main.go
@@ -18,10 +18,7 @@ package main
 
 import (
 	"context"
-	"fmt"
-	"html/template"
 	"log"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/exposure-notifications-server/internal/admin"
@@ -45,7 +42,7 @@ func main() {
 	defer env.Close(ctx)
 
 	router := gin.Default()
-	router.SetFuncMap(funcMap())
+	router.SetFuncMap(admin.TemplateFuncMap)
 	router.LoadHTMLGlob(config.TemplatePath + "/*")
 
 	// Landing page.
@@ -87,99 +84,5 @@ func main() {
 	log.Printf("listening on http://localhost:" + config.Port)
 	if err := router.Run(); err != nil {
 		log.Fatal(err)
-	}
-}
-
-func funcMap() template.FuncMap {
-	return map[string]interface{}{
-		"deref":        deref,
-		"htmlDate":     timestampFormatter("2006-01-02"),
-		"htmlTime":     timestampFormatter("15:04"),
-		"htmlDatetime": timestampFormatter(time.UnixDate),
-	}
-}
-
-func timestampFormatter(f string) func(i interface{}) (string, error) {
-	return func(i interface{}) (string, error) {
-		switch t := i.(type) {
-		case nil:
-			return "", nil
-		case time.Time:
-			if t.IsZero() {
-				return "", nil
-			}
-			return t.UTC().Format(f), nil
-		case *time.Time:
-			if t == nil || t.IsZero() {
-				return "", nil
-			}
-			return t.UTC().Format(f), nil
-		case string:
-			return t, nil
-		default:
-			return "", fmt.Errorf("unknown type %v", t)
-		}
-	}
-}
-
-func deref(i interface{}) (string, error) {
-	switch t := i.(type) {
-	case *string:
-		if t == nil {
-			return "", nil
-		}
-		return *t, nil
-	case *int:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *int8:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *int16:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *int32:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *int64:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *uint:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *uint8:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *uint16:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *uint32:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	case *uint64:
-		if t == nil {
-			return "0", nil
-		}
-		return fmt.Sprintf("%d", *t), nil
-	default:
-		return "", fmt.Errorf("unknown type %T: %v", t, t)
 	}
 }
